moved predict functionality into predict_model
